Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -9,7 +9,7 @@ int welcome() {
     else if(c=='-')
     return a-b;
     else if (c=='/')
-    return a/b;
+    return (int) a/b;
     else if (c=='*')
     return a*b;
 }
